requirements.txt
fastapi
uvicorn
sqlalchemy
pyjwt
redis==4.4.0
psycopg2-binary
asyncpg
pytest
pytest-cov
python-dotenv